refactor: move copy button from turn actions to individual code blocks
ターンフッターのコピーボタンを削除し、各コードブロックにホバーで表示されるコピーボタンを追加
Remove the copy-all button from turn actions footer and add a per-code-block copy button that appears on hover

- Extract useCopyToClipboard hook into shared hooks/ for reuse
- Add CodeBlockWithCopy component wrapping <pre> with a copy button overlay
- Wire CodeBlockWithCopy into markdown renderer via toolMeta
- Clean up TurnActions to only show duration (renders nothing if no duration)
