+++ src/renderer/components/Center/ToolCallGroup/toolMeta.ts
@@ -1,6 +1,7 @@
 import type { ReactNode } from 'react'
 import { createElement } from 'react'
 import { openExternalLink } from '@/lib/openExternalLink'
+import { CodeBlockWithCopy } from '../CodeBlockCopy'
 import {
   IconPencil, IconFile, IconTerminal, IconSearch, IconGlobe,
   IconGitFork, IconChecklist, IconInbox, IconXCircle, IconBook,
@@ -62,7 +63,7 @@ export function LinkRenderer({ href, children }: React.AnchorHTMLAttributes<HTML
   return createElement('a', { href, onClick: handleClick }, children)
 }
 
-export const markdownComponents = { a: LinkRenderer }
+export const markdownComponents = { a: LinkRenderer, pre: CodeBlockWithCopy }
 
 // ── File extension → highlight.js language mapping ────────────────────────────
 // Only includes languages that are registered in ToolCallRow.tsx at module level.
